+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Changed round end message ("%base_players% (%base_scored%)")
 * Fixed POINTS mode wins/loses stats
 * Fixed NPE on fall damage before a player is on any base
+* Fixed block dropping on breaking
 
 ### 1.0.6 Release (13.01.2021 - 18.01.2021)
 * ArenaRegistry will set isDone to false if arena cuboid is not big enough